Use an augeasprovider type as base
