+++ lib/puppet/provider/sshd_config_subsystem/augeas.rb
@@ -3,12 +3,8 @@
 # Copyright (c) 2012 Raphaël Pinson
 # Licensed under the Apache License, Version 2.0
 
-require File.dirname(__FILE__) + '/../../../augeasproviders/provider'
-
-Puppet::Type.type(:sshd_config_subsystem).provide(:augeas) do
+Puppet::Type.type(:sshd_config_subsystem).provide(:augeas, :parent => Puppet::Type.type(:augeasprovider).provider(:default)) do
   desc "Uses Augeas API to update a Subsystem parameter in sshd_config."
-
-  include AugeasProviders::Provider
 
   default_file { '/etc/ssh/sshd_config' }
