+++ lib/puppet/provider/syslog/rsyslog.rb
@@ -3,8 +3,6 @@
 # Copyright (c) 2012 Raphaël Pinson
 # Licensed under the Apache License, Version 2.0
 
-require File.dirname(__FILE__) + '/../../../augeasproviders/provider'
-
 Puppet::Type.type(:syslog).provide(:rsyslog, :parent => :augeas) do
   desc "Uses Augeas API to update an rsyslog.conf entry"
