+++ lib/puppet/type/augeasprovider.rb
@@ -1,0 +1,5 @@
+# Dump augeasprovider type
+
+Puppet::Type.newtype(:augeasprovider) do
+  @doc = 'Dumb Augeas provider type'
+end
