+++ spec/fixtures/unit/puppet/provider/augeasprovider/default/full
